requirements.txt
# Visible Manus - Dependencies

# Core
openai>=2.0.0
python-dotenv>=1.0.0
mcp>=1.0.0
fastmcp>=2.0.0

# Web research
requests>=2.31.0
beautifulsoup4>=4.12.0
duckduckgo-search>=4.0.0

# WebSocket server
websockets>=12.0
